requirements.txt
Django
django-emoji
django-bootstrap4
django-rest-swagger
django-fontawesome
djangorestframework
django-annoying
django-prometheus
django-allauth
django-gravatar2
django-health-check
django-jsonfield

jira
